# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QRX API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors/me/rx-settings:
    get:
      operationId: getMyRxSettings
      tags: [prescriptions]
      summary: Get authenticated doctor's prescription settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorRxSettings"
        "401":
          description: Not authenticated
    put:
      operationId: updateMyRxSettings
      tags: [prescriptions]
      summary: Update authenticated doctor's prescription settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorRxSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorRxSettings"
        "401":
          description: Not authenticated
...
  /doctors/me/subscription/pay/sslcommerz:
    post:
      operationId: payDoctorSubscriptionSslcommerz
      tags: [payments]
      summary: Start an SSLCommerz checkout session for the authenticated doctor's subscription
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                months:
                  type: integer
                  minimum: 1
                  maximum: 120
      responses:
        "200":
          description: Gateway checkout URL
          content:
            application/json:
              schema:
                type: object
                required: [url, tranId]
                properties:
                  url:
                    type: string
                  tranId:
                    type: string
        "400":
          description: Gateway not enabled/configured
        "401":
          description: Not authenticated
        "404":
          description: Not found
  /shop/orders/{id}/pay/sslcommerz:
    post:
      operationId: payShopOrderSslcommerz
      tags: [payments]
      summary: Start an SSLCommerz checkout session for a shop order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gateway checkout URL
          content:
            application/json:
              schema:
                type: object
                required: [url, tranId]
                properties:
                  url:
                    type: string
                  tranId:
                    type: string
        "400":
          description: Gateway not enabled/configured
        "401":
          description: Not authenticated
        "404":
          description: Not found
...
  /reviews:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Submit a patient review (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientReview"
  /doctors/{id}/reviews:
    get:
      operationId: getDoctorReviews
      tags: [reviews]
      summary: Get approved reviews for a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsResponse"
...
  /admin/reviews/{id}/approve:
    post:
      operationId: approveReview
      tags: [admin]
      summary: Approve a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientReview"
  /admin/reviews/{id}/reject:
    post:
      operationId: rejectReview
      tags: [admin]
      summary: Reject/delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ---- EMAIL/SMS LOGS ----
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    DoctorRxSettings:
      type: object
      required: [id, doctorId, pageSize, marginTop, marginRight, marginBottom, marginLeft, showHeader, showQr, showSignature, showFooter]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        headerName:
          type: ["string", "null"]
        headerDegree:
          type: ["string", "null"]
        headerDesignation:
          type: ["string", "null"]
        headerBmdc:
          type: ["string", "null"]
        hospitalName:
          type: ["string", "null"]
        headerAddress:
          type: ["string", "null"]
        headerPhone:
          type: ["string", "null"]
        headerEmail:
          type: ["string", "null"]
        signatureText:
          type: ["string", "null"]
        signatureImage:
          type: ["string", "null"]
        pageSize:
          type: string
        marginTop:
          type: integer
        marginRight:
          type: integer
        marginBottom:
          type: integer
        marginLeft:
          type: integer
        headerHeight:
          type: integer
        footerHeight:
          type: integer
        showHeader:
          type: boolean
        showQr:
          type: boolean
        showSignature:
          type: boolean
        showFooter:
          type: boolean
        footerText:
          type: ["string", "null"]
...
    DoctorRxSettingsInput:
      type: object
      properties:
        headerName:
          type: ["string", "null"]
        headerDegree:
          type: ["string", "null"]
        headerDesignation:
          type: ["string", "null"]
        headerBmdc:
          type: ["string", "null"]
        hospitalName:
          type: ["string", "null"]
        headerAddress:
          type: ["string", "null"]
        headerPhone:
          type: ["string", "null"]
        headerEmail:
          type: ["string", "null"]
        signatureText:
          type: ["string", "null"]
        signatureImage:
          type: ["string", "null"]
        pageSize:
          type: string
        marginTop:
          type: integer
        marginRight:
          type: integer
        marginBottom:
          type: integer
        marginLeft:
          type: integer
        headerHeight:
          type: integer
        footerHeight:
          type: integer
        showHeader:
          type: boolean
        showQr:
          type: boolean
        showSignature:
          type: boolean
        showFooter:
          type: boolean
        footerText:
          type: ["string", "null"]

    # MEDICINES
...
    PatientReview:
      type: object
      required: [id, doctorId, patientName, rating]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: ["string", "null"]
        rating:
          type: integer
        reviewText:
          type: ["string", "null"]
        isApproved:
          type: boolean
        appointmentId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [doctorId, patientName, rating]
      properties:
        doctorId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        rating:
          type: integer
        reviewText:
          type: string
        appointmentId:
          type: integer
    ReviewsResponse:
      type: object
      required: [reviews, avgRating, total]
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/PatientReview"
        avgRating:
          type: number
        total:
          type: integer

    # EMAIL / SMS LOGS
...